Vision/Odom_Tracker/requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
docutils==0.20.1
idna==3.6
Kivy==2.3.0
kivy-deps.angle==0.4.0   
kivy-deps.glew==0.3.1    
kivy-deps.sdl2==0.7.0    
Kivy-Garden==0.1.5       
MouseInfo==0.1.3
pillow==10.2.0
PyAutoGUI==0.9.54        
PyGetWindow==0.0.9       
Pygments==2.17.2
PyMsgBox==1.0.9
pynetworktables==2021.0.0
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.30
pytweening==1.0.7
pywin32==306
requests==2.31.0
urllib3==2.1.0